refactor advisory workspace action routes
